requirements.txt
asgiref==3.4.1
astroid==2.8.0
black==21.9b0
click==8.0.1
Django==3.2.7
django-autoslug==1.9.8
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-flex-fields==0.9.4
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
PyJWT==2.1.0
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
regex==2021.9.24
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
whitenoise==5.3.0
wrapt==1.12.1
